MGS-5849 [#imx-2089] Update header files of OVXLIB and NNRT
Update header files of OVXLIB and NNRT for 6.4.3.p0 RC2 (0717)
release

+++ include/OVXLIB/utils/vsi_nn_util.h
@@ -50,7 +50,7 @@ extern "C" {
 
 #define END_OF_VARIADIC_ARGUMENTS       0xbadcaffe
 #define FOREACH_ARGS(_args, _next, _arg_type) \
-    while(((_arg_type)END_OF_VARIADIC_ARGUMENTS) != (_next = va_arg(_args, _arg_type)))
+    while(((_arg_type)((size_t)END_OF_VARIADIC_ARGUMENTS)) != (_next = va_arg(_args, _arg_type)))
 
 /*-------------------------------------------
                   Functions
